+++ Dyalect/Runtime/BreakException.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace Dyalect.Runtime
+{
+    public sealed class BreakException : Exception
+    {
+    }
+}